[CI] Reorganize entrypoints tests (#27403)
